+++ services/web/src/App.js
@@ -8,7 +8,6 @@ import { useSession } from 'stores';
 
 import Dashboard from 'screens/Dashboard';
 import Docs from 'screens/Docs';
-import Components from 'screens/Components';
 import Invites from 'screens/Invites';
 import NotFound from 'screens/NotFound';
 import Settings from 'screens/Settings';
@@ -46,7 +45,6 @@ const App = () => {
       <Protected path="/users/invites" allowed={Invites} exact />
       <Protected path="/organizations/:id?" allowed={Organizations} />
       <Protected path="/users/:id?" allowed={Users} />
-      <Route path="/docs/ui" component={Components} exact />
       <Route path="/docs/:id?" component={Docs} />
       <Route path="/applications/:id?" component={Applications} />
       <Route path="/logout" component={Logout} exact />
