+++ services/web/src/screens/Docs/Components/index.js
@@ -43,7 +43,7 @@ const options2 = [
 
 @screen
 export default class ComponentsScreen extends React.Component {
-  static layout = 'portal';
+  static layout = null;
 
   contextRef = React.createRef();
