+++ services/web/src/screens/Docs/Docs/table.less
